Generalize clipboard feature for platform independence
The copyToClipboard function would fail unless the platform was Mac OS X
(Or, strictly speaking, if there was no 'pbcopy' binary available). This
patch fixes it so the code runs on Windows. For any other platform it will
now be a noop (which I think is better than having it crash which was the
case before..)